update spline implementation and add extrapolation
two major changes, along with unit tests:

1. Previous spline required two poses before and after the query time.
Then the spline would be fit to those 4 points.
This is inconvenient when wanting to use the spline to extrapolate, because then you're
extrapolating off a very old pose.
This changes uses the first 4 point or last 4 points when at the boundaries.

2. Implement extrapolation which takes the final pose, gets the velocities,
multiplies by dt and calculates the extrapolated pose